# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API documentation for the Translation Management application.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/languages": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "List all languages",
                "operationId": "13f72dfeebea6adfa84684cf31bade15",
                "responses": {
                    "200": {
                        "description": "List of languages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Language"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Languages"
                ],
                "summary": "Create a new language",
                "operationId": "f7b540d1a2e4ba9719a40dc1d8624fa0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "English"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Language created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Language"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List translations with filters",
                "operationId": "ecee4c59cba20fd72cdc6b9c1d8bc084",
                "parameters": [
                    {
                        "name": "language_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tags[]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Translation"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": {
                    "sanctum": {}
                }
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "4589be65f0ce1b409b3bfa17baca71a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "value",
                                    "language_id"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "greeting.hello"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Hello"
                                    },
                                    "language_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        },
        "/api/translations/{id}": {
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update a translation",
                "operationId": "e9759181213b468ce2384c9b16da4abd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "value",
                                    "language_id"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "greeting.hello"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Hello"
                                    },
                                    "language_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    }
                },
                "security": {
                    "sanctum": {}
                }
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation",
                "operationId": "76d12830146dc4acbde69c403b3d95e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Translation deleted"
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        },
        "/api/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations as JSON (filtered by languages and tags)",
                "operationId": "6b6344befbdfe96832018f642d51ce02",
                "parameters": [
                    {
                        "name": "languages[]",
                        "in": "query",
                        "description": "Array of language IDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "tags[]",
                        "in": "query",
                        "description": "Array of tag IDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported translations JSON file",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": {
                    "sanctum": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Language": {
                "title": "Language",
                "required": [
                    "id",
                    "name",
                    "code"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "English"
                    },
                    "code": {
                        "type": "string",
                        "example": "en"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "title": "Tag",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "greeting"
                    }
                },
                "type": "object"
            },
            "Translation": {
                "title": "Translation",
                "required": [
                    "id",
                    "key",
                    "value",
                    "language"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "key": {
                        "type": "string",
                        "example": "greeting.hello"
                    },
                    "value": {
                        "type": "string",
                        "example": "Hello"
                    },
                    "language": {
                        "$ref": "#/components/schemas/Language"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}